requirements.txt
# ReportBench项目依赖包
# 核心数据处理
pandas>=1.3.0
pyyaml>=6.0

# LangChain相关
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# 网络请求和重试
requests>=2.25.0
tenacity>=8.0.0

# 进度条
tqdm>=4.60.0

# 网页解析和抓取
beautifulsoup4>=4.9.0
firecrawl-py>=0.0.1

# 环境变量管理
python-dotenv>=0.19.0

# 其他工具（Python内置，但明确声明版本要求）
# pathlib - Python 3.4+ 内置
# json - Python内置
# re - Python内置
# os - Python内置
# threading - Python内置
# logging - Python内置
# urllib - Python内置
# datetime - Python内置
# concurrent.futures - Python 3.2+ 内置
# hashlib - Python内置
# pickle - Python内置
# string - Python内置
# random - Python内置
# argparse - Python内置
# collections - Python内置
# sys - Python内置
# glob - Python内置
# time - Python内置
# typing - Python 3.5+ 内置（3.9+推荐）

# 注意：此项目需要Python 3.8+